Modify data type

+++ area of rectangle/calculateArea.c
@@ -2,11 +2,11 @@
 //gets area of rectangle
 int main()
 {
-    int length,width;
+    double length,width;
     printf("Enter the length of the rectangle in centimetres: ");
-    scanf("%d",&length);
+    scanf("%lf",&length);
     printf("Enter the width of the rectangle in centimetres: ");
-    scanf("%d",&width);
-    printf("The area of the rectangle is: %d cm\n",length*width);
+    scanf("%lf",&width);
+    printf("The area of the rectangle is: %.1lf cm\n",length*width);
     return 0;
 }
